refactor lua compiler into separate files
